Fixes Rserve connection and -settings

+++ 52n-wps-r/src/main/java/org/n52/wps/server/r/GenericRProcess.java
@@ -100,7 +100,25 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
      */
     private String currentWPSWorkDir;
 
-    private boolean deleteWorkDirectory = false;
+    /**
+     * Indicates if the WPS working directory should be deleted after process
+     * execution
+     */
+    private boolean deleteWPSWorkDirectory = true;
+
+    /**
+     * Indicates if the R working directory should be deleted after process
+     * execution If wpsWorkDirIsRWorkDir is set true, deletion of the directory
+     * shall be determined by deleteWPSWorDirectory
+     * 
+     */
+    private boolean deleteRWorkDirectory = true;
+
+    /**
+     * In case of errors, this variable may be changed to false during runtime
+     * to prevent the system from deleting the wrong files
+     */
+    private boolean temporaryPreventRWorkingDirectoryFromDelete = false;
 
     private RAnnotationParser parser = new RAnnotationParser();
 
@@ -432,7 +450,7 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
         }
 
         // try to delete current local workdir - folder
-        if (this.deleteWorkDirectory) {
+        if (this.deleteWPSWorkDirectory) {
             File workdir = new File(this.currentWPSWorkDir);
             boolean deleted = deleteRecursive(workdir);
             if (!deleted)
@@ -457,8 +475,9 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
      *            does not call open- or close-operations.
      * @throws REXPMismatchException
      * @throws RserveException
+     * @throws ExceptionReport
      */
-    private void setRWorkingDirectoryBeforeProcessing(RConnection rCon) throws REXPMismatchException, RserveException
+    private void setRWorkingDirectoryBeforeProcessing(RConnection rCon) throws REXPMismatchException, RserveException, ExceptionReport
     {
         R_Config rconf = R_Config.getInstance();
 
@@ -483,6 +502,7 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
                 // R starts from a work directory dependent on the behaviour and
                 // configuration of the R/Rserve
                 // installation
+                this.wpsWorkDirIsRWorkDir = false;
                 String randomFolderName = "wps4r-r-workdir-" + UUID.randomUUID().toString().substring(0, 8);
                 rCon.eval("dir.create(\"" + randomFolderName + "\")"); // quotation
                                                                        // marks!
@@ -496,17 +516,26 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
 
         } else if (config_RWorkDir.trim().equalsIgnoreCase("preset")) {
             // setting the R working directory relative to default R directory
+            this.wpsWorkDirIsRWorkDir = false;
             String randomFolderName = "wps4r-r-workdir-" + UUID.randomUUID().toString().substring(0, 8);
             rCon.eval("dir.create(\"" + randomFolderName + "\")"); // quotation
                                                                    // marks!
             result = rCon.eval("setwd(\"" + randomFolderName + "\")"); // quotation
-                                                                       // marks!
-        } else if (config_RWorkDir.trim().equalsIgnoreCase("temporary")) {
+        }                                                              // marks!
+        // setting the R working directory in a temporal folder
+        else if (config_RWorkDir.trim().equalsIgnoreCase("temporary")) {
             if (isLocalhost) {
                 this.wpsWorkDirIsRWorkDir = true;
                 result = rCon.eval("setwd(\"" + this.currentWPSWorkDir.replace("\\", "/") + "\")");
             } else {
-                result = rCon.eval("setwd(\"tempdir()\")");
+                this.wpsWorkDirIsRWorkDir = false;
+                try {
+                    result = rCon.eval("setwd(\"tempdir()\")");
+                } catch (RserveException e) {
+                    temporaryPreventRWorkingDirectoryFromDelete = true;
+                    throw new ExceptionReport("Invalid configuration of WPS4R, failed to create temporal working directory", ExceptionReport.REMOTE_COMPUTATION_ERROR, e);
+                }
+
             }
         } else {
 
@@ -542,6 +571,13 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
                 rconf.setConfigVariable(RWPSConfigVariables.R_WORK_DIR, "default");
                 setRWorkingDirectoryBeforeProcessing(rCon);
             }
+        }
+
+
+        if (rCon.eval("length(dir()) == 0").asInteger() == 1) {
+            temporaryPreventRWorkingDirectoryFromDelete = true;
+            throw new ExceptionReport("Non-empty R working directory on process startup. The process will not be executed to prevent the system from damage."
+                    + "Please reconsider the configuration of WPS4R", ExceptionReport.REMOTE_COMPUTATION_ERROR);
         }
 
         log.debug("[R] Old wd: " + result.asString() + " | New wd: " + rCon.eval("getwd()").asString());
@@ -588,9 +624,12 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
             connection.eval("setwd(\"" + r_basedir + "\")");
             // should be true usually, if not, workdirectory has been
             // changed unexpectedly (prob. inside script)
-            if (currentwd != r_basedir) {
-                log.debug("[R] unlinking (recursive) " + currentwd);
-                connection.eval("unlink(\"" + currentwd + "\", recursive=TRUE)");
+            if (currentwd != r_basedir && deleteRWorkDirectory) {
+                if (!temporaryPreventRWorkingDirectoryFromDelete) {
+                    log.debug("[R] unlinking (recursive) " + currentwd);
+                    connection.eval("unlink(\"" + currentwd + "\", recursive=TRUE)");
+                } else
+                    temporaryPreventRWorkingDirectoryFromDelete = false;
             } else
                 log.warn("Unexpected R workdirectory at end of R session, check the R sript for unwanted workdirectory changes");
         }
@@ -972,7 +1011,7 @@ public class GenericRProcess extends AbstractObservableAlgorithm {
             }
         }
 
-        Class[] easyLiterals =
+        Class<?>[] easyLiterals =
                 new Class[] { LiteralByteBinding.class, LiteralDoubleBinding.class, LiteralFloatBinding.class, LiteralIntBinding.class, LiteralLongBinding.class, LiteralShortBinding.class,
                         LiteralStringBinding.class };
